+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [0.17.15] - 2022-07-11
+- Fix width in csv data table
 - Fix save button in event modal
 
 ## [0.17.14] - 2022-07-10